requirements.txt

django-admin-autocomplete-filter==0.6.1
django-admin-rangefilter==0.7.0
django-admin-sortable==2.2.3
django-bootstrap-form==3.4
django-material-admin==1.7.16
django-multiselectfield==0.1.12
django-search-admin-autocomplete==0.2.0